requirements.txt
absl-py==1.2.0
astunparse==1.6.3
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-default-group==1.2.2
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==1.12
fonttools==4.37.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.0
gunicorn==20.1.0
h5py==3.7.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.2.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.22.4
missingno==0.5.1
multimethod==1.9
networkx==2.8.6
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandas-profiling==3.2.0
phik==0.12.2
Pillow==9.2.0
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
tangled-up-in-unicode==0.2.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
termcolor==2.0.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
urwid==2.1.2
valuedispatch==0.0.1
visions==0.7.4
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
